Support component story format
This updates story loading to support the Component Story Format (CSF).  It also updates the glob to scan for in `.mdx` files.  The `storyOf` format is still supported, so this is likely not a breaking change.

The current method of loading stories does not support the component story format or MDX.